split SoaCuspCorrectionBasisSet.h file.
