+++ src/QMCWaveFunctions/LCAO/LCAOrbitalBuilder.cpp
@@ -18,18 +18,17 @@
 #include "LCAOrbitalBuilder.h"
 #include "OhmmsData/AttributeSet.h"
 #include "QMCWaveFunctions/SPOSet.h"
-#include "QMCWaveFunctions/LCAO/MultiQuinticSpline1D.h"
-#include "QMCWaveFunctions/LCAO/SoaCartesianTensor.h"
-#include "QMCWaveFunctions/LCAO/SoaSphericalTensor.h"
-#include "QMCWaveFunctions/LCAO/SoaAtomicBasisSet.h"
-#include "QMCWaveFunctions/LCAO/SoaLocalizedBasisSet.h"
-#include "QMCWaveFunctions/LCAO/LCAOrbitalSet.h"
-//#include "QMCWaveFunctions/LCAO/RadialOrbitalSetBuilder.h"
-#include "QMCWaveFunctions/LCAO/AOBasisBuilder.h"
-#include "QMCWaveFunctions/LCAO/MultiFunctorAdapter.h"
+#include "MultiQuinticSpline1D.h"
+#include "SoaCartesianTensor.h"
+#include "SoaSphericalTensor.h"
+#include "SoaAtomicBasisSet.h"
+#include "SoaLocalizedBasisSet.h"
+#include "LCAOrbitalSet.h"
+#include "AOBasisBuilder.h"
+#include "MultiFunctorAdapter.h"
 #if !defined(QMC_COMPLEX)
-#include "QMCWaveFunctions/LCAO/LCAOrbitalSetWithCorrection.h"
-#include "QMCWaveFunctions/LCAO/CuspCorrectionConstruction.h"
+#include "LCAOrbitalSetWithCorrection.h"
+#include "CuspCorrectionConstruction.h"
 #endif
 #include "hdf/hdf_archive.h"
 #include "Message/CommOperators.h"
