+++ src/QMCWaveFunctions/LCAO/SoaCuspCorrectionBasisSet.h
@@ -20,7 +20,7 @@
 #include "Configuration.h"
 #include "QMCWaveFunctions/BasisSetBase.h"
 #include "Particle/DistanceTableData.h"
-#include "QMCWaveFunctions/LCAO/MultiQuinticSpline1D.h"
+#include "MultiQuinticSpline1D.h"
 
 namespace qmcplusplus
 {
@@ -101,226 +101,6 @@ public:
       du_z[j] -= dphi[i] * dr[2] / r;
       d2u[j] += d2phi[i] + 2 * dphi[i] / r;
     }
-  }
-};
-
-/** A localized basis set derived from BasisSetBase<typename COT::ValueType>
- *
- * This class performs the evaluation of the basis functions and their
- * derivatives for each of the N-particles in a configuration.
- * The template parameter COT denotes Centered-Orbital-Type which provides
- * a set of localized orbitals associated with a center.
- */
-struct SoaCuspCorrection
-{
-  typedef QMCTraits::ValueType ValueType;
-  typedef QMCTraits::RealType RealType;
-  typedef VectorSoaContainer<ValueType, 5> VGLVector_t;
-  typedef SPOSet::ValueMatrix_t ValueMatrix_t;
-  typedef SPOSet::GradMatrix_t GradMatrix_t;
-  typedef SPOSet::GradVector_t GradVector_t;
-  typedef SPOSet::ValueVector_t ValueVector_t;
-  typedef ParticleSet::PosType PosType;
-
-  ///number of centers, e.g., ions
-  size_t NumCenters;
-  ///number of quantum particles
-  size_t NumTargets;
-  ///number of quantum particles
-  const int myTableIndex;
-  ///size of the basis set
-  int BasisSetSize;
-
-  ///COMPLEX WON'T WORK
-  typedef CuspCorrectionAtomicBasis<RealType> COT;
-
-  int unused = 1;
-  /** container of the unique pointers to the Atomic Orbitals
-   *
-   * size of LOBasisSet = number of centers (atoms)
-   * should use unique_ptr once COT is fixed for better performance
-   */
-  std::vector<std::shared_ptr<const COT>> LOBasisSet;
-
-  Matrix<RealType> myVGL;
-
-  /** constructor
-   * @param ions ionic system
-   * @param els electronic system
-   */
-  SoaCuspCorrection(ParticleSet& ions, ParticleSet& els) : myTableIndex(els.addTable(ions))
-  {
-    NumCenters = ions.getTotalNum();
-    NumTargets = els.getTotalNum();
-    LOBasisSet.resize(NumCenters);
-  }
-
-  /** copy constructor */
-  SoaCuspCorrection(const SoaCuspCorrection& a) = default;
-
-
-  /** set BasisSetSize and allocate mVGL container
-   */
-  void setBasisSetSize(int nbs)
-  {
-    BasisSetSize = nbs;
-    //THIS NEEDS TO BE FIXE for OpenMP
-    myVGL.resize(5, BasisSetSize);
-  }
-
-  /** compute VGL
-   * @param P quantum particleset
-   * @param iat active particle
-   * @param vgl Matrix(5,BasisSetSize)
-   * @param trialMove if true, use getTempDists()/getTempDispls()
-   */
-  inline void evaluateVGL(const ParticleSet& P, int iat, VGLVector_t& vgl)
-  {
-    myVGL = 0.0;
-
-    const auto& d_table = P.getDistTable(myTableIndex);
-    const auto& dist    = (P.activePtcl == iat) ? d_table.getTempDists() : d_table.getDistRow(iat);
-    const auto& displ   = (P.activePtcl == iat) ? d_table.getTempDispls() : d_table.getDisplRow(iat);
-    for (int c = 0; c < NumCenters; c++)
-    {
-      if (LOBasisSet[c])
-      {
-        LOBasisSet[c]->evaluate_vgl(dist[c], displ[c], myVGL[0], myVGL[1], myVGL[2], myVGL[3], myVGL[4]);
-      }
-    }
-
-    {
-      const auto v_in  = myVGL[0];
-      const auto gx_in = myVGL[1];
-      const auto gy_in = myVGL[2];
-      const auto gz_in = myVGL[3];
-      const auto l_in  = myVGL[4];
-      auto v_out       = vgl.data(0);
-      auto gx_out      = vgl.data(1);
-      auto gy_out      = vgl.data(2);
-      auto gz_out      = vgl.data(3);
-      auto l_out       = vgl.data(4);
-      for (size_t i = 0; i < BasisSetSize; ++i)
-      {
-        v_out[i] += v_in[i];
-        gx_out[i] += gx_in[i];
-        gy_out[i] += gy_in[i];
-        gz_out[i] += gz_in[i];
-        l_out[i] += l_in[i];
-      }
-    }
-  }
-
-  inline void evaluate_vgl(const ParticleSet& P, int iat, ValueVector_t& psi, GradVector_t& dpsi, ValueVector_t& d2psi)
-  {
-    myVGL = 0.0;
-
-    const auto& d_table = P.getDistTable(myTableIndex);
-    const auto& dist    = (P.activePtcl == iat) ? d_table.getTempDists() : d_table.getDistRow(iat);
-    const auto& displ   = (P.activePtcl == iat) ? d_table.getTempDispls() : d_table.getDisplRow(iat);
-    for (int c = 0; c < NumCenters; c++)
-    {
-      if (LOBasisSet[c])
-      {
-        LOBasisSet[c]->evaluate_vgl(dist[c], displ[c], myVGL[0], myVGL[1], myVGL[2], myVGL[3], myVGL[4]);
-      }
-    }
-
-    const auto v_in  = myVGL[0];
-    const auto gx_in = myVGL[1];
-    const auto gy_in = myVGL[2];
-    const auto gz_in = myVGL[3];
-    const auto l_in  = myVGL[4];
-    for (size_t i = 0; i < BasisSetSize; ++i)
-    {
-      psi[i] += v_in[i];
-      dpsi[i][0] += gx_in[i];
-      dpsi[i][1] += gy_in[i];
-      dpsi[i][2] += gz_in[i];
-      d2psi[i] += l_in[i];
-    }
-  }
-
-  inline void evaluate_vgl(const ParticleSet& P,
-                           int iat,
-                           int idx,
-                           ValueMatrix_t& psi,
-                           GradMatrix_t& dpsi,
-                           ValueMatrix_t& d2psi)
-  {
-    myVGL = 0.0;
-
-    const auto& d_table = P.getDistTable(myTableIndex);
-    const auto& dist    = (P.activePtcl == iat) ? d_table.getTempDists() : d_table.getDistRow(iat);
-    const auto& displ   = (P.activePtcl == iat) ? d_table.getTempDispls() : d_table.getDisplRow(iat);
-    for (int c = 0; c < NumCenters; c++)
-    {
-      if (LOBasisSet[c])
-      {
-        LOBasisSet[c]->evaluate_vgl(dist[c], displ[c], myVGL[0], myVGL[1], myVGL[2], myVGL[3], myVGL[4]);
-      }
-    }
-
-    const auto v_in  = myVGL[0];
-    const auto gx_in = myVGL[1];
-    const auto gy_in = myVGL[2];
-    const auto gz_in = myVGL[3];
-    const auto l_in  = myVGL[4];
-    for (size_t i = 0; i < BasisSetSize; ++i)
-    {
-      psi[idx][i] += v_in[i];
-      dpsi[idx][i][0] += gx_in[i];
-      dpsi[idx][i][1] += gy_in[i];
-      dpsi[idx][i][2] += gz_in[i];
-      d2psi[idx][i] += l_in[i];
-    }
-  }
-
-  /** compute values for the iat-paricle move
-   *
-   * Always uses getTempDists() and getTempDispls()
-   */
-  inline void evaluateV(const ParticleSet& P, int iat, ValueType* restrict vals)
-  {
-    ValueType* tmp_vals = myVGL[0];
-
-    std::fill_n(tmp_vals, myVGL.size(), 0.0);
-
-    const auto& d_table = P.getDistTable(myTableIndex);
-    const auto& dist    = (P.activePtcl == iat) ? d_table.getTempDists() : d_table.getDistRow(iat);
-
-    //THIS IS SERIAL, only way to avoid this is to use myVGL
-    for (int c = 0; c < NumCenters; c++)
-    {
-      if (LOBasisSet[c])
-      {
-        LOBasisSet[c]->evaluate(dist[c], tmp_vals);
-      }
-    }
-
-    { //collect
-      const auto v_in = myVGL[0];
-      for (size_t i = 0; i < BasisSetSize; ++i)
-      {
-        vals[i] += v_in[i];
-      }
-    }
-  }
-
-  /** add a new set of Centered Atomic Orbitals
-   * @param icenter the index of the center
-   * @param aos a set of Centered Atomic Orbitals
-   */
-  void add(int icenter, std::unique_ptr<COT> aos) { LOBasisSet[icenter].reset(aos.release()); }
-  void addVGL(const ParticleSet& P, int iat, VGLVector_t& vgl) { evaluateVGL(P, iat, vgl); }
-  void addV(const ParticleSet& P, int iat, ValueType* restrict vals) { evaluateV(P, iat, vals); }
-  void add_vgl(const ParticleSet& P, int iat, int idx, ValueMatrix_t& vals, GradMatrix_t& dpsi, ValueMatrix_t& d2psi)
-  {
-    evaluate_vgl(P, iat, idx, vals, dpsi, d2psi);
-  }
-  void add_vector_vgl(const ParticleSet& P, int iat, ValueVector_t& vals, GradVector_t& dpsi, ValueVector_t& d2psi)
-  {
-    evaluate_vgl(P, iat, vals, dpsi, d2psi);
   }
 };
 } // namespace qmcplusplus
